Added dummy driver and logs
